+++ src/main/kotlin/org/danilopianini/plagiarismdetector/Test.kt
