+++ src/main/kotlin/org/danilopianini/plagiarismdetector/provider/criteria/GitHubSearchCriteria.kt
@@ -1,4 +1,4 @@
-package provider.criteria
+package org.danilopianini.plagiarismdetector.provider.criteria
 
 private const val SPACE_CHAR = " "
