+++ src/main/kotlin/org/danilopianini/plagiarismdetector/repository/Repository.kt
@@ -1,4 +1,4 @@
-package repository
+package org.danilopianini.plagiarismdetector.repository
 
 import java.io.File
