+++ src/main/kotlin/org/danilopianini/plagiarismdetector/repository/content/RepoContentSupplierCloneStrategy.kt
@@ -1,4 +1,4 @@
-package repository.content
+package org.danilopianini.plagiarismdetector.repository.content
 
 import org.apache.commons.io.FileUtils
 import org.apache.commons.io.filefilter.DirectoryFileFilter
