+++ src/test/kotlin/org/danilopianini/plagiarismdetector/logger/LoggerTest.kt
@@ -1,4 +1,4 @@
-package logger
+package org.danilopianini.plagiarismdetector.logger
 
 import ch.qos.logback.classic.Level
 import ch.qos.logback.classic.Logger
